+++ GgafDxCore/src/jp/ggaf/dxcore/model/GgafDxBoardSetModel.cpp
@@ -183,36 +183,36 @@ void GgafDxBoardSetModel::restore() {
         //1pxあたりのuvの大きさを求める
 //        float tex_width  = (float)(model_pTextureConnection->peek()->_pD3DXIMAGE_INFO->Width); //テクスチャの幅(px)
 //        float tex_height = (float)(model_pTextureConnection->peek()->_pD3DXIMAGE_INFO->Height); //テクスチャの高さ(px)
-        float du = 0.0f; //1.0 / tex_width  / 100000.0; //テクスチャの幅1pxの10000分の1px
-        float dv = 0.0f; //1.0 / tex_height / 100000.0; //テクスチャの高さ1pxの10000分の1px
+        double du = 0.0; //1.0 / tex_width  / 100000.0; //テクスチャの幅1pxの10000分の1px
+        double dv = 0.0; //1.0 / tex_height / 100000.0; //テクスチャの高さ1pxの10000分の1px
         for (int i = 0; i < _set_num; i++) {
             //左上
             paVertex[i*4 + 0].x = 0.0f;
             paVertex[i*4 + 0].y = 0.0f;
             paVertex[i*4 + 0].z = 0.0f;
-            paVertex[i*4 + 0].tu = du;
-            paVertex[i*4 + 0].tv = dv;
+            paVertex[i*4 + 0].tu = (float)du;
+            paVertex[i*4 + 0].tv = (float)dv;
             paVertex[i*4 + 0].index = (float)i;
             //右上
             paVertex[i*4 + 1].x = xdata.width;
             paVertex[i*4 + 1].y = 0.0f;
             paVertex[i*4 + 1].z = 0.0f;
-            paVertex[i*4 + 1].tu = (1.0 / xdata.col_texture_split) - du;
-            paVertex[i*4 + 1].tv = dv;
+            paVertex[i*4 + 1].tu = (float)((1.0 / xdata.col_texture_split) - du);
+            paVertex[i*4 + 1].tv = (float)dv;
             paVertex[i*4 + 1].index = (float)i;
             //左下
             paVertex[i*4 + 2].x = 0.0f;
             paVertex[i*4 + 2].y = xdata.height;
             paVertex[i*4 + 2].z = 0.0f;
-            paVertex[i*4 + 2].tu = du;
-            paVertex[i*4 + 2].tv = (1.0 / xdata.row_texture_split) - dv;
+            paVertex[i*4 + 2].tu = (float)du;
+            paVertex[i*4 + 2].tv = (float)((1.0 / xdata.row_texture_split) - dv);
             paVertex[i*4 + 2].index = (float)i;
             //右下
             paVertex[i*4 + 3].x = xdata.width;
             paVertex[i*4 + 3].y = xdata.height;
             paVertex[i*4 + 3].z = 0.0f;
-            paVertex[i*4 + 3].tu = (1.0 / xdata.col_texture_split) - du;
-            paVertex[i*4 + 3].tv = (1.0 / xdata.row_texture_split) - dv;
+            paVertex[i*4 + 3].tu = (float)((1.0 / xdata.col_texture_split) - du);
+            paVertex[i*4 + 3].tv = (float)((1.0 / xdata.row_texture_split) - dv);
             paVertex[i*4 + 3].index = (float)i;
          }
